@@ -10,6 +10,8 @@ import org.jmqtt.broker.cluster.ClusterMessageTransfer;
 import org.jmqtt.broker.cluster.ClusterSessionManager;
 import org.jmqtt.broker.cluster.DefaultClusterMessageTransfer;
 import org.jmqtt.broker.cluster.DefaultClusterSessionManager;
+import org.jmqtt.broker.cluster.redis.RedisClusterMessageTransfer;
+import org.jmqtt.broker.cluster.redis.RedisClusterSessionManager;
 import org.jmqtt.broker.dispatcher.DefaultDispatcherMessage;
 import org.jmqtt.broker.dispatcher.MessageDispatcher;
 import org.jmqtt.broker.processor.*;
@@ -24,11 +26,13 @@ import org.jmqtt.common.helper.MixAll;
 import org.jmqtt.common.helper.RejectHandler;
 import org.jmqtt.common.helper.ThreadFactoryImpl;
 import org.jmqtt.common.log.LoggerName;
+import org.jmqtt.manage.HttpServer;
 import org.jmqtt.remoting.netty.ChannelEventListener;
 import org.jmqtt.remoting.netty.NettyRemotingServer;
 import org.jmqtt.remoting.netty.RequestProcessor;
 import org.jmqtt.store.*;
 import org.jmqtt.store.memory.DefaultMqttStore;
+import org.jmqtt.store.redis.RedisMqttStore;
 import org.jmqtt.store.rocksdb.RDBMqttStore;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -42,34 +46,35 @@ public class BrokerController {
 
     private static final Logger log = LoggerFactory.getLogger(LoggerName.BROKER);
 
-    private BrokerConfig          brokerConfig;
-    private NettyConfig           nettyConfig;
+    private BrokerConfig brokerConfig;
+    private NettyConfig nettyConfig;
     private StoreConfig storeConfig;
     private ClusterConfig clusterConfig;
-    private ExecutorService      connectExecutor;
-    private ExecutorService      pubExecutor;
-    private ExecutorService       subExecutor;
-    private ExecutorService       pingExecutor;
-    private LinkedBlockingQueue   connectQueue;
-    private LinkedBlockingQueue   pubQueue;
-    private LinkedBlockingQueue   subQueue;
-    private LinkedBlockingQueue   pingQueue;
-    private ChannelEventListener  channelEventListener;
-    private NettyRemotingServer   remotingServer;
-    private MessageDispatcher     messageDispatcher;
-    private FlowMessageStore      flowMessageStore;
-    private SubscriptionMatcher   subscriptionMatcher;
-    private WillMessageStore      willMessageStore;
-    private RetainMessageStore    retainMessageStore;
-    private OfflineMessageStore   offlineMessageStore;
-    private SubscriptionStore     subscriptionStore;
-    private SessionStore          sessionStore;
-    private AbstractMqttStore     abstractMqttStore;
-    private ConnectPermission     connectPermission;
-    private PubSubPermission      pubSubPermission;
-    private ReSendMessageService  reSendMessageService;
+    private ExecutorService connectExecutor;
+    private ExecutorService pubExecutor;
+    private ExecutorService subExecutor;
+    private ExecutorService pingExecutor;
+    private LinkedBlockingQueue<Runnable> connectQueue;
+    private LinkedBlockingQueue<Runnable> pubQueue;
+    private LinkedBlockingQueue<Runnable> subQueue;
+    private LinkedBlockingQueue<Runnable> pingQueue;
+    private ChannelEventListener channelEventListener;
+    private NettyRemotingServer remotingServer;
+    private MessageDispatcher messageDispatcher;
+    private FlowMessageStore flowMessageStore;
+    private SubscriptionMatcher subscriptionMatcher;
+    private WillMessageStore willMessageStore;
+    private RetainMessageStore retainMessageStore;
+    private OfflineMessageStore offlineMessageStore;
+    private SubscriptionStore subscriptionStore;
+    private SessionStore sessionStore;
+    private AbstractMqttStore abstractMqttStore;
+    private ConnectPermission connectPermission;
+    private PubSubPermission pubSubPermission;
+    private ReSendMessageService reSendMessageService;
     private ClusterSessionManager clusterSessionManager;
     private ClusterMessageTransfer clusterMessageTransfer;
+    private HttpServer httpServer;
 
 
     public BrokerController(BrokerConfig brokerConfig, NettyConfig nettyConfig, StoreConfig storeConfig, ClusterConfig clusterConfig) {
@@ -78,22 +83,22 @@ public class BrokerController {
         this.storeConfig = storeConfig;
         this.clusterConfig = clusterConfig;
 
-        this.connectQueue = new LinkedBlockingQueue(100000);
-        this.pubQueue = new LinkedBlockingQueue(100000);
-        this.subQueue = new LinkedBlockingQueue(100000);
-        this.pingQueue = new LinkedBlockingQueue(10000);
+        this.connectQueue = new LinkedBlockingQueue<>(100000);
+        this.pubQueue = new LinkedBlockingQueue<>(100000);
+        this.subQueue = new LinkedBlockingQueue<>(100000);
+        this.pingQueue = new LinkedBlockingQueue<>(10000);
 
-        {//store pluggable
+        {
+            //store pluggable
             switch (storeConfig.getStoreType()) {
                 case 1:
                     this.abstractMqttStore = new RDBMqttStore(storeConfig);
-                    this.clusterSessionManager = new DefaultClusterSessionManager();
-                    this.clusterMessageTransfer = new DefaultClusterMessageTransfer();
                     break;
-                default:
+                case 2:
+                    this.abstractMqttStore = new RedisMqttStore(clusterConfig);
+                    break;
+                case 3:
                     this.abstractMqttStore = new DefaultMqttStore();
-                    this.clusterSessionManager = new DefaultClusterSessionManager();
-                    this.clusterMessageTransfer = new DefaultClusterMessageTransfer();
                     break;
             }
             try {
@@ -110,7 +115,8 @@ public class BrokerController {
             this.sessionStore = this.abstractMqttStore.getSessionStore();
         }
 
-        {// permission pluggable
+        {
+            // permission pluggable
             this.connectPermission = new DefaultConnectPermission();
             this.pubSubPermission = new DefaultPubSubPermission();
         }
@@ -120,6 +126,7 @@ public class BrokerController {
 
         this.channelEventListener = new ClientLifeCycleHookService(willMessageStore, messageDispatcher);
         this.remotingServer = new NettyRemotingServer(brokerConfig, nettyConfig, channelEventListener);
+        this.httpServer = new HttpServer(nettyConfig);
         this.reSendMessageService = new ReSendMessageService(offlineMessageStore, flowMessageStore);
 
         int coreThreadNum = Runtime.getRuntime().availableProcessors();
@@ -151,6 +158,36 @@ public class BrokerController {
                 pingQueue,
                 new ThreadFactoryImpl("PingThread"),
                 new RejectHandler("heartbeat", 100000));
+
+
+        {
+            if (checkClusterMode()) {
+                // cluster
+                switch (clusterConfig.getClusterComponentName()) {
+                    case "local":
+                        this.clusterSessionManager = new DefaultClusterSessionManager(sessionStore, subscriptionStore);
+                        this.clusterMessageTransfer = new DefaultClusterMessageTransfer(messageDispatcher, clusterConfig);
+                        break;
+                    case "redis":
+                        this.clusterSessionManager = new RedisClusterSessionManager(sessionStore, subscriptionStore);
+                        this.clusterMessageTransfer = new RedisClusterMessageTransfer(messageDispatcher, (RedisMqttStore) abstractMqttStore);
+                        break;
+                }
+            }
+        }
+
+    }
+
+    /**
+     * 校验集群模式
+     *
+     * @return
+     */
+    private boolean checkClusterMode() {
+        if (clusterConfig != null) {
+            return "cluster".equals(clusterConfig.getMode());
+        }
+        return false;
     }
 
 
@@ -161,7 +198,8 @@ public class BrokerController {
         MixAll.printProperties(log, storeConfig);
         MixAll.printProperties(log, clusterConfig);
 
-        {//init and register mqtt remoting processor
+        {
+            //init and register mqtt remoting processor
             RequestProcessor connectProcessor = new ConnectProcessor(this);
             RequestProcessor disconnectProcessor = new DisconnectProcessor(this);
             RequestProcessor pingProcessor = new PingProcessor();
@@ -185,21 +223,58 @@ public class BrokerController {
             this.remotingServer.registerProcessor(MqttMessageType.PUBCOMP, pubCompProcessor, subExecutor);
         }
 
-        this.messageDispatcher.start();
-        this.reSendMessageService.start();
-        this.remotingServer.start();
+        if (this.messageDispatcher != null) {
+            this.messageDispatcher.start();
+        }
+        if (this.reSendMessageService != null) {
+            this.reSendMessageService.start();
+        }
+        if (this.remotingServer != null) {
+            this.remotingServer.start();
+        }
+        if (this.clusterSessionManager != null) {
+            this.clusterSessionManager.startup();
+        }
+        if (this.clusterMessageTransfer != null) {
+            this.clusterMessageTransfer.startup();
+        }
+        if (this.httpServer != null) {
+            this.httpServer.start();
+        }
         log.info("JMqtt Server start success and version = {}", brokerConfig.getVersion());
     }
 
     public void shutdown() {
-        this.remotingServer.shutdown();
-        this.connectExecutor.shutdown();
-        this.pubExecutor.shutdown();
-        this.subExecutor.shutdown();
-        this.pingExecutor.shutdown();
-        this.messageDispatcher.shutdown();
-        this.reSendMessageService.shutdown();
-        this.abstractMqttStore.shutdown();
+        if (this.remotingServer != null) {
+            this.remotingServer.shutdown();
+        }
+        if (this.connectExecutor != null) {
+            this.connectExecutor.shutdown();
+        }
+        if (this.pubExecutor != null) {
+            this.pubExecutor.shutdown();
+        }
+        if (this.subExecutor != null) {
+            this.subExecutor.shutdown();
+        }
+        if (this.pingExecutor != null) {
+            this.pingExecutor.shutdown();
+        }
+        if (this.messageDispatcher != null) {
+            this.messageDispatcher.shutdown();
+        }
+        if (this.reSendMessageService != null) {
+            this.reSendMessageService.shutdown();
+        }
+        if (this.abstractMqttStore != null) {
+            this.abstractMqttStore.shutdown();
+        }
+        if (this.clusterMessageTransfer != null) {
+            this.clusterMessageTransfer.shutdown();
+        }
+        if (this.clusterSessionManager != null) {
+            this.clusterSessionManager.shutdown();
+        }
     }
 
     public BrokerConfig getBrokerConfig() {
@@ -226,19 +301,19 @@ public class BrokerController {
         return pingExecutor;
     }
 
-    public LinkedBlockingQueue getConnectQueue() {
+    public LinkedBlockingQueue<Runnable> getConnectQueue() {
         return connectQueue;
     }
 
-    public LinkedBlockingQueue getPubQueue() {
+    public LinkedBlockingQueue<Runnable> getPubQueue() {
         return pubQueue;
     }
 
-    public LinkedBlockingQueue getSubQueue() {
+    public LinkedBlockingQueue<Runnable> getSubQueue() {
         return subQueue;
     }
 
-    public LinkedBlockingQueue getPingQueue() {
+    public LinkedBlockingQueue<Runnable> getPingQueue() {
         return pingQueue;
     }
 
