@@ -6,13 +6,13 @@ package org.jmqtt.broker.acl;
 public interface PubSubPermission {
 
     /**
-     * verfy the clientId whether can publish message to the topic
+     * verify the clientId whether can publish message to the topic
      */
-    boolean publishVerfy(String clientId,String topic);
+    boolean publishVerify(String clientId,String topic);
 
     /**
-     * verfy the clientId whether can subscribe the topic
+     * verify the clientId whether can subscribe the topic
      */
-    boolean subscribeVerfy(String clientId,String topic);
+    boolean subscribeVerify(String clientId,String topic);
 
 }
