@@ -5,12 +5,12 @@ import org.jmqtt.broker.acl.PubSubPermission;
 public class DefaultPubSubPermission implements PubSubPermission {
 
     @Override
-    public boolean publishVerfy(String clientId, String topic) {
+    public boolean publishVerify(String clientId, String topic) {
         return true;
     }
 
     @Override
-    public boolean subscribeVerfy(String clientId, String topic) {
+    public boolean subscribeVerify(String clientId, String topic) {
         return true;
     }
 }
