@@ -3,6 +3,7 @@ package org.jmqtt.broker.processor;
 import org.jmqtt.broker.cluster.ClusterMessageTransfer;
 import org.jmqtt.broker.cluster.command.CommandCode;
 import org.jmqtt.broker.cluster.command.CommandReqOrResp;
+import org.jmqtt.broker.cluster.redis.RedisClusterMessageTransfer;
 import org.jmqtt.broker.dispatcher.MessageDispatcher;
 import org.jmqtt.common.model.Message;
 import org.jmqtt.common.model.MessageHeader;
@@ -10,18 +11,25 @@ import org.jmqtt.store.RetainMessageStore;
 
 public abstract class AbstractMessageProcessor {
 
-    private MessageDispatcher    messageDispatcher;
-    private RetainMessageStore   retainMessageStore;
+    private MessageDispatcher messageDispatcher;
+    private RetainMessageStore retainMessageStore;
     private ClusterMessageTransfer clusterMessageTransfer;
 
-    public AbstractMessageProcessor(MessageDispatcher messageDispatcher, RetainMessageStore retainMessageStore,ClusterMessageTransfer clusterMessageTransfer) {
+    public AbstractMessageProcessor(MessageDispatcher messageDispatcher, RetainMessageStore retainMessageStore, ClusterMessageTransfer clusterMessageTransfer) {
         this.messageDispatcher = messageDispatcher;
         this.retainMessageStore = retainMessageStore;
         this.clusterMessageTransfer = clusterMessageTransfer;
     }
 
     protected void processMessage(Message message) {
-        this.messageDispatcher.appendMessage(message);
+        /**
+         * @author yance
+         * @date 2020.10.10
+         * <p> 生产者生产一条消息,消费者会收到两条消息,解决方法,如果是集群模式,不需要向本地增加当前消息 <p/>
+         */
+        if (null == clusterMessageTransfer) {
+            this.messageDispatcher.appendMessage(message);
+        }
         boolean retain = (boolean) message.getHeader(MessageHeader.RETAIN);
         if (retain) {
             int qos = (int) message.getHeader(MessageHeader.QOS);
@@ -38,8 +46,10 @@ public abstract class AbstractMessageProcessor {
     }
 
     private void dispatcherMessage2Cluster(Message message) {
-        CommandReqOrResp commandReqOrResp = new CommandReqOrResp(CommandCode.MESSAGE_CLUSTER_TRANSFER,message);
-        clusterMessageTransfer.sendMessage(commandReqOrResp);
+        CommandReqOrResp commandReqOrResp = new CommandReqOrResp(CommandCode.MESSAGE_CLUSTER_TRANSFER, message);
+        if (clusterMessageTransfer != null) {
+            clusterMessageTransfer.sendMessage(commandReqOrResp);
+        }
     }
 
 }
