@@ -49,7 +49,7 @@ public class ConnectProcessor implements RequestProcessor {
     private ClusterSessionManager clusterSessionManager;
 
 
-    public ConnectProcessor(BrokerController brokerController){
+    public ConnectProcessor(BrokerController brokerController) {
         this.flowMessageStore = brokerController.getFlowMessageStore();
         this.willMessageStore = brokerController.getWillMessageStore();
         this.offlineMessageStore = brokerController.getOfflineMessageStore();
@@ -63,112 +63,115 @@ public class ConnectProcessor implements RequestProcessor {
 
     @Override
     public void processRequest(ChannelHandlerContext ctx, MqttMessage mqttMessage) {
-        MqttConnectMessage connectMessage = (MqttConnectMessage)mqttMessage;
+        MqttConnectMessage connectMessage = (MqttConnectMessage) mqttMessage;
         MqttConnectReturnCode returnCode = null;
         int mqttVersion = connectMessage.variableHeader().version();
         String clientId = connectMessage.payload().clientIdentifier();
-        boolean cleansession = connectMessage.variableHeader().isCleanSession();
+        boolean cleanSession = connectMessage.variableHeader().isCleanSession();
         String userName = connectMessage.payload().userName();
         byte[] password = connectMessage.payload().passwordInBytes();
         ClientSession clientSession = null;
         boolean sessionPresent = false;
-        try{
-            if(!versionValid(mqttVersion)){
+        try {
+            if (!versionValid(mqttVersion)) {
                 returnCode = MqttConnectReturnCode.CONNECTION_REFUSED_UNACCEPTABLE_PROTOCOL_VERSION;
-            } else if(!clientIdVerfy(clientId)){
+            } else if (!clientIdVerify(clientId)) {
                 returnCode = MqttConnectReturnCode.CONNECTION_REFUSED_IDENTIFIER_REJECTED;
-            } else if(onBlackList(RemotingHelper.getRemoteAddr(ctx.channel()),clientId)){
+            } else if (onBlackList(RemotingHelper.getRemoteAddr(ctx.channel()), clientId)) {
                 returnCode = MqttConnectReturnCode.CONNECTION_REFUSED_NOT_AUTHORIZED;
-            } else if(!authentication(clientId,userName,password)){
+            } else if (!authentication(clientId, userName, password)) {
                 returnCode = MqttConnectReturnCode.CONNECTION_REFUSED_BAD_USER_NAME_OR_PASSWORD;
-            } else{
+            } else {
                 int heartbeatSec = connectMessage.variableHeader().keepAliveTimeSeconds();
-                if(!keepAlive(clientId,ctx,heartbeatSec)){
-                    log.warn("[CONNECT] -> set heartbeat failure,clientId:{},heartbeatSec:{}",clientId,heartbeatSec);
+                if (!keepAlive(clientId, ctx, heartbeatSec)) {
+                    log.warn("[CONNECT] -> set heartbeat failure,clientId:{},heartbeatSec:{}", clientId, heartbeatSec);
                     throw new Exception("set heartbeat failure");
                 }
-                CommandReqOrResp response = clusterSessionManager.process(new CommandReqOrResp(CommandCode.CONNECT_QUERY_LAST_STATE));
-                if(Objects.nonNull(response.getBody()) && BooleanUtils.toBoolean(response.getBody().toString())){
+                CommandReqOrResp response = null;
+                if (this.clusterSessionManager != null) {
+                    response = clusterSessionManager.process(new CommandReqOrResp(CommandCode.CONNECT_QUERY_LAST_STATE));
+                }
+                if (Objects.nonNull(response) && Objects.nonNull(response.getBody()) && BooleanUtils.toBoolean(response.getBody().toString())) {
                     ClientSession previousClient = ConnectManager.getInstance().getClient(clientId);
-                    if(previousClient != null){
+                    if (previousClient != null) {
                         previousClient.getCtx().close();
                         ConnectManager.getInstance().removeClient(clientId);
                     }
                 }
-                if(cleansession){
-                    clientSession = createNewClientSession(clientId,ctx);
+                if (cleanSession) {
+                    clientSession = createNewClientSession(clientId, ctx);
                     sessionPresent = false;
-                }else{
-                    if(Objects.nonNull(response.getBody())){
-                        clientSession = reloadClientSession(ctx,clientId);
+                } else {
+                    if (Objects.nonNull(response) && Objects.nonNull(response.getBody())) {
+                        clientSession = reloadClientSession(ctx, clientId);
                         sessionPresent = true;
-                    }else{
-                        clientSession = new ClientSession(clientId,false,ctx);
+                    } else {
+                        clientSession = new ClientSession(clientId, false, ctx);
                         sessionPresent = false;
                     }
                 }
-                sessionStore.setSession(clientId,true);
+                sessionStore.setSession(clientId, true);
                 boolean willFlag = connectMessage.variableHeader().isWillFlag();
-                if(willFlag){
+                if (willFlag) {
                     boolean willRetain = connectMessage.variableHeader().isWillRetain();
                     int willQos = connectMessage.variableHeader().willQos();
                     String willTopic = connectMessage.payload().willTopic();
                     byte[] willPayload = connectMessage.payload().willMessageInBytes();
-                    storeWillMsg(clientId,willRetain,willQos,willTopic,willPayload);
+                    storeWillMsg(clientId, willRetain, willQos, willTopic, willPayload);
                 }
                 returnCode = MqttConnectReturnCode.CONNECTION_ACCEPTED;
-                NettyUtil.setClientId(ctx.channel(),clientId);
-                ConnectManager.getInstance().putClient(clientId,clientSession);
+                NettyUtil.setClientId(ctx.channel(), clientId);
+                ConnectManager.getInstance().putClient(clientId, clientSession);
             }
-            MqttConnAckMessage ackMessage = MessageUtil.getConnectAckMessage(returnCode,sessionPresent);
+            MqttConnAckMessage ackMessage = MessageUtil.getConnectAckMessage(returnCode, sessionPresent);
             ctx.writeAndFlush(ackMessage);
-            if(returnCode != MqttConnectReturnCode.CONNECTION_ACCEPTED){
+            if (returnCode != MqttConnectReturnCode.CONNECTION_ACCEPTED) {
                 ctx.close();
-                log.warn("[CONNECT] -> {} connect failure,returnCode={}",clientId,returnCode);
+                log.warn("[CONNECT] -> {} connect failure,returnCode={}", clientId, returnCode);
                 return;
             }
-            log.info("[CONNECT] -> {} connect to this mqtt server",clientId);
+            log.info("[CONNECT] -> {} connect to this mqtt server", clientId);
             reConnect2SendMessage(clientId);
             newClientNotify(clientSession);
-        }catch(Exception ex){
-            log.warn("[CONNECT] -> Service Unavailable: cause={}",ex);
+        } catch (Exception ex) {
+            log.warn("[CONNECT] -> Service Unavailable: cause={}", ex);
             returnCode = MqttConnectReturnCode.CONNECTION_REFUSED_SERVER_UNAVAILABLE;
-            MqttConnAckMessage ackMessage = MessageUtil.getConnectAckMessage(returnCode,sessionPresent);
+            MqttConnAckMessage ackMessage = MessageUtil.getConnectAckMessage(returnCode, sessionPresent);
             ctx.writeAndFlush(ackMessage);
             ctx.close();
         }
     }
 
-    private void newClientNotify(ClientSession clientSession){
+    private void newClientNotify(ClientSession clientSession) {
 
     }
-    
-    private boolean keepAlive(String clientId,ChannelHandlerContext ctx,int heatbeatSec){
-        if(this.connectPermission.verfyHeartbeatTime(clientId,heatbeatSec)){
-            int keepAlive = (int)(heatbeatSec * 1.5f);
-            if(ctx.pipeline().names().contains("idleStateHandler")){
+
+    private boolean keepAlive(String clientId, ChannelHandlerContext ctx, int heatbeatSec) {
+        if (this.connectPermission.verifyHeartbeatTime(clientId, heatbeatSec)) {
+            int keepAlive = (int) (heatbeatSec * 1.5f);
+            if (ctx.pipeline().names().contains("idleStateHandler")) {
                 ctx.pipeline().remove("idleStateHandler");
             }
-            ctx.pipeline().addFirst("idleStateHandler",new IdleStateHandler(keepAlive,0,0));
+            ctx.pipeline().addFirst("idleStateHandler", new IdleStateHandler(keepAlive, 0, 0));
             return true;
         }
         return false;
     }
 
-    private void storeWillMsg(String clientId,boolean willRetain,int willQos,String willTopic,byte[] willPayload){
-        Map<String,Object> headers = new HashMap<>();
-        headers.put(MessageHeader.RETAIN,willRetain);
-        headers.put(MessageHeader.QOS,willQos);
-        headers.put(MessageHeader.TOPIC,willTopic);
-        headers.put(MessageHeader.WILL,true);
-        Message message = new Message(Message.Type.WILL,headers,willPayload);
+    private void storeWillMsg(String clientId, boolean willRetain, int willQos, String willTopic, byte[] willPayload) {
+        Map<String, Object> headers = new HashMap<>();
+        headers.put(MessageHeader.RETAIN, willRetain);
+        headers.put(MessageHeader.QOS, willQos);
+        headers.put(MessageHeader.TOPIC, willTopic);
+        headers.put(MessageHeader.WILL, true);
+        Message message = new Message(Message.Type.WILL, headers, willPayload);
         message.setClientId(clientId);
-        willMessageStore.storeWillMessage(clientId,message);
-        log.info("[WillMessageStore] : {} store will message:{}",clientId,message);
+        willMessageStore.storeWillMessage(clientId, message);
+        log.info("[WillMessageStore] : {} store will message:{}", clientId, message);
     }
 
-    private ClientSession createNewClientSession(String clientId,ChannelHandlerContext ctx){
-        ClientSession clientSession = new ClientSession(clientId,true);
+    private ClientSession createNewClientSession(String clientId, ChannelHandlerContext ctx) {
+        ClientSession clientSession = new ClientSession(clientId, true);
         clientSession.setCtx(ctx);
         //clear previous sessions
         this.flowMessageStore.clearClientFlowCache(clientId);
@@ -181,36 +184,36 @@ public class ConnectProcessor implements RequestProcessor {
     /**
      * cleansession is false, reload client session
      */
-    private ClientSession reloadClientSession(ChannelHandlerContext ctx,String clientId){
-            ClientSession clientSession = new ClientSession(clientId,false);
-            clientSession.setCtx(ctx);
-            CommandReqOrResp response = clusterSessionManager.process(new CommandReqOrResp(CommandCode.CONNECT_GET_SUBSCRIPTIONS,clientId));
-            Collection<Subscription> subscriptions = (Collection<Subscription>) response.getBody();
-            for(Subscription subscription : subscriptions){
-                this.subscriptionMatcher.subscribe(subscription);
-            }
-            return clientSession;
+    private ClientSession reloadClientSession(ChannelHandlerContext ctx, String clientId) {
+        ClientSession clientSession = new ClientSession(clientId, false);
+        clientSession.setCtx(ctx);
+        CommandReqOrResp response = clusterSessionManager.process(new CommandReqOrResp(CommandCode.CONNECT_GET_SUBSCRIPTIONS, clientId));
+        Collection<Subscription> subscriptions = (Collection<Subscription>) response.getBody();
+        for (Subscription subscription : subscriptions) {
+            this.subscriptionMatcher.subscribe(subscription);
+        }
+        return clientSession;
     }
 
-    private void reConnect2SendMessage(String clientId){
+    private void reConnect2SendMessage(String clientId) {
         this.reSendMessageService.put(clientId);
         this.reSendMessageService.wakeUp();
     }
 
-    private boolean authentication(String clientId,String username,byte[] password){
-        return this.connectPermission.authentication(clientId,username,password);
+    private boolean authentication(String clientId, String username, byte[] password) {
+        return this.connectPermission.authentication(clientId, username, password);
     }
 
-    private boolean onBlackList(String remoteAddr,String clientId){
-        return this.connectPermission.onBlacklist(remoteAddr,clientId);
+    private boolean onBlackList(String remoteAddr, String clientId) {
+        return this.connectPermission.onBlacklist(remoteAddr, clientId);
     }
 
-    private boolean clientIdVerfy(String clientId){
-        return this.connectPermission.clientIdVerfy(clientId);
+    private boolean clientIdVerify(String clientId) {
+        return this.connectPermission.clientIdVerify(clientId);
     }
 
-    private boolean versionValid(int mqttVersion){
-        if(mqttVersion == 3 || mqttVersion == 4){
+    private boolean versionValid(int mqttVersion) {
+        if (mqttVersion == 3 || mqttVersion == 4) {
             return true;
         }
         return false;
