@@ -1,7 +1,5 @@
 package org.jmqtt.broker.processor;
 
-import io.netty.channel.ChannelFuture;
-import io.netty.channel.ChannelFutureListener;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.handler.codec.mqtt.MqttMessage;
 import org.jmqtt.common.log.LoggerName;
@@ -17,7 +15,7 @@ public class PubAckProcessor implements RequestProcessor {
     private Logger log = LoggerFactory.getLogger(LoggerName.MESSAGE_TRACE);
     private FlowMessageStore flowMessageStore;
 
-    public PubAckProcessor(FlowMessageStore flowMessageStore){
+    public PubAckProcessor(FlowMessageStore flowMessageStore) {
         this.flowMessageStore = flowMessageStore;
     }
 
@@ -25,11 +23,11 @@ public class PubAckProcessor implements RequestProcessor {
     public void processRequest(ChannelHandlerContext ctx, MqttMessage mqttMessage) {
         String clientId = NettyUtil.getClientId(ctx.channel());
         int messageId = MessageUtil.getMessageId(mqttMessage);
-        log.debug("[PubAck] -> Recieve PubAck message,clientId={},msgId={}",clientId,messageId);
-        if(!flowMessageStore.containSendMsg(clientId,messageId)){
-            log.warn("[PubAck] -> The message is not cached in Flow,clientId={},msgId={}",clientId,messageId);
+        log.info("[PubAck] -> Recieve PubAck message,clientId={},msgId={}", clientId, messageId);
+        if (!flowMessageStore.containSendMsg(clientId, messageId)) {
+            log.warn("[PubAck] -> The message is not cached in Flow,clientId={},msgId={}", clientId, messageId);
             return;
         }
-        flowMessageStore.releaseSendMsg(clientId,messageId);
+        flowMessageStore.releaseSendMsg(clientId, messageId);
     }
 }
