@@ -5,10 +5,10 @@ import org.eclipse.paho.client.mqttv3.persist.MemoryPersistence;
 
 public class Consumer {
     private static final String broker = "tcp://127.0.0.1:1883";
-    private static final String topic = "MQTT/+";
+    private static final String topic = "MQTT/TOPIC";
     private static final String clientId = "MQTT_SUB_CLIENT";
 
-    public static void main(String[] args) throws MqttException, InterruptedException {
+    public static void main(String[] args) throws MqttException {
         MqttClient subClient = getMqttClient();
         subClient.setCallback(new MqttCallback() {
             @Override
@@ -17,25 +17,26 @@ public class Consumer {
             }
 
             @Override
-            public void messageArrived(String s, MqttMessage mqttMessage) throws Exception {
+            public void messageArrived(String s, MqttMessage mqttMessage) {
                 System.out.println("From topic: " + s);
                 System.out.println("Message content: " + new String(mqttMessage.getPayload()));
             }
 
             @Override
             public void deliveryComplete(IMqttDeliveryToken iMqttDeliveryToken) {
-
+                System.out.println("deliveryComplete");
             }
+
         });
         subClient.subscribe(topic);
     }
 
 
-    private static MqttClient getMqttClient(){
+    private static MqttClient getMqttClient() {
         try {
-            MqttClient pubClient = new MqttClient(broker,clientId,new MemoryPersistence());
+            MqttClient pubClient = new MqttClient(broker, clientId, new MemoryPersistence());
             MqttConnectOptions connectOptions = new MqttConnectOptions();
-            connectOptions.setCleanSession(true);
+            connectOptions.setCleanSession(false);
             System.out.println("Connecting to broker: " + broker);
             pubClient.connect(connectOptions);
             return pubClient;
